requirements.txt
opencv-python>=4.7.0
numpy>=1.24.0
scipy>=1.10.0
requests>=2.28.0
pillow>=9.5.0
flask>=2.3.0
flask-cors>=3.0.10
platformdirs>=3.10
inference  
